Create LocationsAdder transformer
