+++ subway_pipeline.py
@@ -1,10 +1,10 @@
 import pandas as pd
 import os
-from collections import defaultdict
 
 from actuals_adder import ActualsAdder
 from destinations_adder import DestinationsAdder
 from location_id_translator import LocationIdTranslator
+from locations_adder import LocationsAdder
 
 
 class SubwayPipeline():
@@ -24,18 +24,20 @@ class SubwayPipeline():
 
     def load(self):
         vehicle_datapoints = self._load_vehicle_datapoints()
-        destinations_adder = DestinationsAdder(self.locations_path)
-        results_with_destinations = destinations_adder.fit_transform(
+
+        locations_adder = LocationsAdder(self.locations_path)
+        datapoints_with_locations = locations_adder.fit_transform(
             vehicle_datapoints
         )
+
+        destinations_adder = DestinationsAdder(self.locations_path)
+        results_with_destinations = destinations_adder.fit_transform(
+            datapoints_with_locations
+        )
+
         actuals_adder = ActualsAdder(self.actuals_path)
         datapoints = actuals_adder.fit_transform(results_with_destinations)
         return datapoints.dropna()
-
-    # Return array of all location IDs in lexicographical order
-    def _all_locs_sorted(self):
-        loc_frame = pd.read_csv(self.locations_path)
-        return sorted(loc_frame["loc_id"].__array__())
 
     # Build a dataframe mapping each pattern ID to the corresponding terminal
     # stop GTFS ID
@@ -105,66 +107,4 @@ class SubwayPipeline():
             ["terminal_stop_id", "timestamp_y", "pattern_id"],
             axis=1
         ).rename({"timestamp_x": "timestamp"}, axis="columns")
-
-        n_hot_locations = self._n_hot_vehicle_locations_by_generation()
-
-        generations = frame_with_terminal_modes["generation"].to_list()
-
-        n_hot_per_row_list_of_lists = list(
-            map(
-                lambda generation: n_hot_locations[generation], generations
-            )
-        )
-        n_hot_per_row = pd.DataFrame(
-            n_hot_per_row_list_of_lists,
-            columns=self._all_locs_sorted()
-        )
-        return frame_with_terminal_modes.join(n_hot_per_row, how="inner")
-
-    # Builds a map of generations to the list of n-hot encoded vehicle
-    # positions for that generation.
-    def _n_hot_vehicle_locations_by_generation(self):
-        raw_frame = pd.read_csv(
-            self.vehicles_path,
-            usecols=[
-                'current_location_id',
-                'generation',
-                'gtfs_trip_id',
-                'length_of_time_at_current_location',
-                'ocs_trip_id',
-                'offset_departure_seconds_from_now',
-                'pattern_id',
-                'timestamp',
-                'vehicle_id'
-            ]
-        )
-        grouped_frame = raw_frame.groupby(by="generation")
-
-        occupied_locations_by_generation = {}
-        generational_n_hot_locations = {}
-
-        for generation in grouped_frame:
-            generation_id = generation[0]
-            locations = generation[1]["current_location_id"]
-
-            occupied_locations_by_generation[generation_id] = set()
-            for location in locations:
-                occupied_locations_by_generation[generation_id].add(location)
-
-            generational_n_hot_locations[generation_id] = defaultdict(
-                lambda: 0
-            )
-            for location in occupied_locations_by_generation[generation_id]:
-                generational_n_hot_locations[generation_id].update(
-                    {location: 1}
-                )
-
-        generational_n_hot_location_lists = {}
-        for generation in generational_n_hot_locations:
-            location_map = generational_n_hot_locations[generation]
-            n_hot_columns = []
-            for location in self._all_locs_sorted():
-                n_hot_columns.append(location_map[location])
-                generational_n_hot_location_lists[generation] = n_hot_columns
-
-        return generational_n_hot_location_lists
+        return frame_with_terminal_modes
